LineDebounce moved to separate file and template added
