+++ Firmware/targets/arm/Inc/hardwareConfig.h
@@ -2,6 +2,7 @@
 #define TARGETS_ARM_INC_HARDWARECONFIG_H_
 
 #define ENCODER_DEBOUNCE_STABLE 5
+#define BUTTON_DEBOUNCE_STABLE 20
 #define ENCODER_PRESC 4
 
 #endif  // TARGETS_ARM_INC_HARDWARECONFIG_H_
